Refactor profile screen associated files
