+++ src/components/main_screens/profile/options/TradeHistoryScreen.jsx
@@ -7,11 +7,11 @@ import {
   Text,
   View,
 } from "react-native";
-import { TRADE_HISTORY_FETCH_AMOUNT } from "../../../constants/ApiConstants";
-import AlgoquantApiContext from "../../../constants/ApiContext";
-import { THEME } from "../../../constants/Theme";
-import { jobHistoryColumns } from "../../../helpers/tableColumns";
-import CustomTable from "../../reusable_components/CustomTable";
+import { TRADE_HISTORY_FETCH_AMOUNT } from "../../../../constants/ApiConstants";
+import AlgoquantApiContext from "../../../../constants/ApiContext";
+import { THEME } from "../../../../constants/Theme";
+import { jobHistoryColumns } from "../../../../helpers/tableColumns";
+import CustomTable from "../../../reusable_components/CustomTable";
 
 export default function TradeHistoryScreen() {
   const [history, setHistory] = useState([]);
